# Conflicts:
#	src/main/java/com/example/carservice/config/MongoConfig.java
#	src/main/java/com/example/carservice/model/Car.java
#	src/main/java/com/example/carservice/service/CarService.java

+++ src/main/java/com/example/carservice/DTO/CarDTO.java
@@ -24,7 +24,7 @@ public class CarDTO implements Serializable {
 
     private String currency;
 
-    private BigDecimal dailyRate;
+    private Double dailyRate;
 
     @Override
     public String toString() {
